This commit adds codebase for [feature request #526](https://github.com/rstudio/blogdown/issues/526#issue-757702574).
Added features:
1. function `serve_site` (within the `R/serve.R` file) has gained a new argument `.site_dir = NULL` after the existing arguments.
This argument lets the user specify the directory containing their website, or a container for this project.
The argument is passed in the call to `serve`. The addition of the argument behind the existing `...` argument should ensure that this does not break existing users code base.
(affected file lines: 41, 54)
2. The description for `serve_site` (in `R/serve.R`) has gained a description for the new argument `.site_dir` with the following description:
@param .site_dir Directory to search for site configuration file (affected file lines: 30)
3. The function `serve_it` (in `R/serve.R`) now returns a function with argument `.site_dir = NULL`.
This argument is passed from `serve_site` and is passed on to the function `site_root`.
No documentation is added, as this function has no existing documentation.
(affected lines: 91, 92)
4. The function `site_root` (in `R/utils.R`) has similarly gained the same argument `.site_dir = NULL`.
If the argument is provided with the default `NULL` value, this argument has no effect.
Otherwise the argument is assumed to be the directory for the site to be served,
and thus the start for the search for the site configuration files (`config.toml/config.yaml`).
In effect, if it is provided the directory is changed after a call to `on.exit(setwd(owd), add = TRUE)`
and thus the directory is changed to the current working directory which was the also the behaviour after this change.
A comment has been added to describe the behaviour within the source code.
(affected lines: 361, 364, 365)
5. A new contributor has been added to the `DESCRIPTION` file.
`person("Oliver", "Madsen", role = "ctb", email = "oliver.p.madsen@gmail.com")`