replace Even with Jane (#710)
It seems that Even is no longer maintained. Even has not been updated since Jan 19, 2021, and many issues and pull requests were marked as stale and closed because the author did not deal with them. It might be difficult for users to get help. In contrast, Jane is actively maintained.